requirements.txt
# --------- pytorch --------- #
#torch==2.0.0+cpu
#torchvision==0.15.0+cpu

# --------- gpt --------- #
tiktoken==0.4.0
protobuf==3.20.*

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0

# --------- deployment --------- #
gradio
boto3